Generalize VaspInputAgent into PeriodicDFTAgent (scale-aware base)
Step 1 of the simulation-orchestrator routing refactor.

Architectural commitment: nothing in source code is hardcoded to a
specific simulation software. Each scale (classical MD, MLIP,
periodic DFT, future: molecular DFT, ...) has one scale-aware agent
class; software-specific behavior lives in skill bundles under
scilink/skills/<scale>/<software>/. MDSimulationAgent + MLIPAgent
were already in this shape; VaspInputAgent was the holdout.

This commit does the smallest move that fits VaspInputAgent into
the pattern:

  - Rename file:  scilink/agents/sim_agents/vasp_agent.py
              ->  scilink/agents/sim_agents/periodic_dft_agent.py
              via git mv (history preserved).

  - Rename class: VaspInputAgent -> PeriodicDFTAgent.
                  Declare SKILL_DOMAIN = "periodic_dft" and
                  SUPPORTED_SOFTWARE = ("vasp",) class vars so the
                  orchestrator can discover capabilities.

  - Move skill bundle:
      scilink/skills/vasp/vasp_input_generation/vasp_input_generation.md
   -> scilink/skills/periodic_dft/vasp/vasp.md
      Matches the MD agent's layout (skills/<scale>/<software>/<software>.md).

  - Backward compat:
      A thin scilink/agents/sim_agents/vasp_agent.py re-exports
      VaspInputAgent = PeriodicDFTAgent so existing callers
      (simulation_orchestrator_tools, vasp_quality, vasp_updater,
      user scripts) keep working.
      An internal _LEGACY_SKILL_ALIASES map silently rewrites the
      old "vasp_input_generation" skill name to "vasp" at load
      time, so the default in generate_vasp_inputs(skill=...) still
      resolves.

  - Re-exports: scilink.agents.sim_agents now exposes both
      PeriodicDFTAgent (new canonical) and VaspInputAgent (alias).

The agent's INTERNAL implementation still has VASP-flavored prompt
content (INCAR/POSCAR/KPOINTS terminology threaded through the
prompts under instruct.VASP_INPUT_GENERATION_INSTRUCTIONS). That's
phase-2 work -- extract software-specific prose into the per-engine
skill bundle so the agent class itself is fully software-agnostic.
Phase 1 only changes the EXTERNAL API to be scale-aware; nothing
in the rest of the codebase breaks.

Verification:
  - PeriodicDFTAgent imports clean; SKILL_DOMAIN and
    SUPPORTED_SOFTWARE present.
  - VaspInputAgent IS PeriodicDFTAgent (same class object via alias).
  - load_skill('vasp', domain='periodic_dft') returns the six
    canonical sections.
  - Legacy skill name 'vasp_input_generation' resolves via the
    alias map -- existing callers in vasp_quality / vasp_updater
    don't need updates.
  - tests/test_skill_loader.py: 9/9 pass.

Next commits: capability discovery API across all scale agents,
user-available-software config, orchestrator router tool.
